add comments for tab view setup
