--- browse/browse.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from "@angular/core";
-
-@Component({
-    selector: "Browse",
-    moduleId: module.id,
-    templateUrl: "./browse.component.html"
-})
-export class BrowseComponent implements OnInit {
-    constructor() {
-
-    }
-
-    ngOnInit(): void {
-
-    }
-}
