--- home/home.component.html
@@ -1,2 +1,0 @@
-<GridLayout class="page appPage homePage">
-</GridLayout>
