--- home/home.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from "@angular/core";
-
-@Component({
-    selector: "Home",
-    moduleId: module.id,
-    templateUrl: "./home.component.html"
-})
-export class HomeComponent implements OnInit {
-    constructor() {
-
-    }
-
-    ngOnInit(): void {
-
-    }
-}
