--- search/search.component.html
@@ -1,2 +1,0 @@
-<GridLayout class="page appPage searchPage">
-</GridLayout>
