--- search/search.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from "@angular/core";
-
-@Component({
-    selector: "Search",
-    moduleId: module.id,
-    templateUrl: "./search.component.html"
-})
-export class SearchComponent implements OnInit {
-    constructor() {
-
-    }
-
-    ngOnInit(): void {
-
-    }
-}
