+++ tabs/browse/browse.component.html
@@ -1,0 +1,6 @@
+<!--
+This is the component view that represents the 'Browse' tab.
+Feel free to customize layouts and components to change how the tab view looks.
+-->
+<GridLayout class="page appPage browsePage">
+</GridLayout>
