+++ tabs/browse/browse.component.ts
@@ -1,0 +1,20 @@
+import { Component, OnInit } from "@angular/core";
+
+@Component({
+    selector: "Browse",
+    moduleId: module.id,
+    templateUrl: "./browse.component.html"
+})
+export class BrowseComponent implements OnInit {
+    constructor() {
+        /* ***********************************************************
+        * Use the constructor to inject services.
+        *************************************************************/
+    }
+
+    ngOnInit(): void {
+        /* ***********************************************************
+        * Use the "ngOnInit" handler to initialize data for the view.
+        *************************************************************/
+    }
+}
