+++ tabs/featured/featured.component.ts
@@ -1,0 +1,20 @@
+import { Component, OnInit, ViewChild } from "@angular/core";
+
+@Component({
+    selector: "Featured",
+    moduleId: module.id,
+    templateUrl: "./featured.component.html"
+})
+export class FeaturedComponent implements OnInit {
+    constructor() {
+        /* ***********************************************************
+        * Use the constructor to inject services.
+        *************************************************************/
+    }
+
+    ngOnInit(): void {
+        /* ***********************************************************
+        * Use the "ngOnInit" handler to initialize data for the view.
+        *************************************************************/
+    }
+}
