+++ tabs/home/home.component.ts
@@ -1,0 +1,20 @@
+import { Component, OnInit } from "@angular/core";
+
+@Component({
+    selector: "Home",
+    moduleId: module.id,
+    templateUrl: "./home.component.html"
+})
+export class HomeComponent implements OnInit {
+    constructor() {
+        /* ***********************************************************
+        * Use the constructor to inject services.
+        *************************************************************/
+    }
+
+    ngOnInit(): void {
+        /* ***********************************************************
+        * Use the "ngOnInit" handler to initialize data for the view.
+        *************************************************************/
+    }
+}
