+++ tabs/search/search.component.ts
@@ -1,0 +1,20 @@
+import { Component, OnInit } from "@angular/core";
+
+@Component({
+    selector: "Search",
+    moduleId: module.id,
+    templateUrl: "./search.component.html"
+})
+export class SearchComponent implements OnInit {
+    constructor() {
+        /* ***********************************************************
+        * Use the constructor to inject services.
+        *************************************************************/
+    }
+
+    ngOnInit(): void {
+        /* ***********************************************************
+        * Use the "ngOnInit" handler to initialize data for the view.
+        *************************************************************/
+    }
+}
