+++ tabs/settings/settings.component.html
@@ -1,0 +1,6 @@
+<!--
+This is the component view that represents the 'Settings' tab.
+Feel free to customize layouts and components to change how the tab view looks.
+-->
+<GridLayout class="page appPage settingsPage">
+</GridLayout>
