+++ tabs/settings/settings.component.ts
@@ -1,0 +1,20 @@
+import { Component, OnInit, ViewChild } from "@angular/core";
+
+@Component({
+    selector: "Settings",
+    moduleId: module.id,
+    templateUrl: "./settings.component.html"
+})
+export class SettingsComponent implements OnInit {
+    constructor() {
+        /* ***********************************************************
+        * Use the constructor to inject services.
+        *************************************************************/
+    }
+
+    ngOnInit(): void {
+        /* ***********************************************************
+        * Use the "ngOnInit" handler to initialize data for the view.
+        *************************************************************/
+    }
+}
