+++ tabs/tabs.component.ts
@@ -9,13 +9,24 @@ import { isAndroid } from "platform";
 })
 export class TabsComponent implements OnInit {
     constructor() {
-
+        /* ***********************************************************
+        * Use the constructor to inject app services that will be needed for
+        * the whole tab navigation layout as a whole.
+        *************************************************************/
     }
 
     ngOnInit(): void {
-
+        /* ***********************************************************
+        * Use the "ngOnInit" handler to initialize data for the whole tab
+        * navigation layout as a whole.
+        *************************************************************/
     }
 
+    /* ***********************************************************
+    * The "getIconSource" function returns the correct tab icon source
+    * depending on whether the app is ran on Android or iOS.
+    * You can find all resources in /App_Resources/os
+    *************************************************************/
     getIconSource(icon: string): string {
         return isAndroid ? "res://tab_" + icon : "res://tabIcons/tab_" + icon;
     }
